wayang_classifier_web/requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3
scipy==1.11.4
Werkzeug==2.3.7
